+++ NEWS.md
@@ -2,6 +2,9 @@
 
 See the ReadMe for the latest two events
 
+
+* 2021-09-13 aMSE 0.0.0.075 Implemented the option of having an initial depletion as well as application of historical catches.
+
 * 2021-09-10 aMSE 0.0.0.100 Many small, but significant, changes. Modified getdata which interacts with the ###HS.R file and added the
 zoneDP$NAS data input. Added a summary recdev plot to condition tab.